Bruk provider for notifikasjoner
Ved å bruke provideren for notifikasjoner forholdsvis høyt
oppe i DOM-en, så vil det ikke forsvinne i en kort periode når
man bytter side.
